# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API da Cozinha",
    "version": "1.0.0",
    "description": "Documentação da API de pedidos da cozinha"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/pedidos": {
      "get": {
        "summary": "Lista todos os pedidos",
        "responses": {
          "200": {
            "description": "Lista de pedidos retornada com sucesso"
          }
        }
      },
      "post": {
        "summary": "Cria um novo pedido",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prato": {
                    "type": "string",
                    "example": "Macarronada"
                  }
                },
                "required": ["prato"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido criado com sucesso"
          },
          "400": {
            "description": "Erro de validação"
          }
        }
      }
    },

    "/menu": {
      "get": {
        "summary": "Retorna o cardápio",
        "responses": {
          "200": {
            "description": "Lista de itens do cardápio"
          }
        }
      }
    },

    "/pedidos/{id}": {
      "get": {
        "summary": "Busca pedido por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido encontrado"
          },
          "404": {
            "description": "Pedido não encontrado"
          },
          "400": {
            "description": "ID inválido"
          }
        }
      },

      "put": {
        "summary": "Atualiza o status do pedido",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "pronto"
                  }
                },
                "required": ["status"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pedido atualizado"
          },
          "404": {
            "description": "Pedido não encontrado"
          },
          "400": {
            "description": "Erro de validação"
          }
        }
      },

      "delete": {
        "summary": "Deleta um pedido",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Pedido deletado com sucesso"
          },
          "404": {
            "description": "Pedido não encontrado"
          }
        }
      }
    },

    "/pedidos/status/{status}": {
      "get": {
        "summary": "Filtra pedidos por status",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "preparando"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedidos filtrados"
          },
          "404": {
            "description": "Nenhum pedido encontrado com esse status"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    }
  }
}